Renamed  to  as described in point 7 of the following [pull request](https://github.com/Baekalfen/PyBoy/pull/22)
